@@ -1,6 +1,5 @@
 package com.test.beercatalogue.controller;
 
-import com.test.beercatalogue.model.Beer;
 import com.test.beercatalogue.model.Manufacturer;
 import com.test.beercatalogue.repository.ManufacturerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,7 +10,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
